--- src/main/java/Counter.java
@@ -1,11 +1,0 @@
-public class Counter {
-    private int counter = 0;
-
-    public int getCounter() {
-        return counter;
-    }
-
-    public void sumUp() {
-        counter = counter + 1;
-    }
-}
