+++ src/main/java/multithreads/MyRunnable.java
@@ -1,0 +1,16 @@
+package multithreads;
+
+public class MyRunnable implements Runnable {
+    private Counter counter;
+
+    MyRunnable(Counter counter) {
+        this.counter = counter;
+    }
+
+    @Override
+    public void run() {
+        while (counter.getCounter() < Counter.END_COUNT) {
+            counter.sumUp();
+        }
+    }
+}
